+++ src/features/listing/ui/ListingCard.tsx
@@ -1,33 +1,15 @@
 'use client';
 
 import ListingCardTag from './ListingCardTag';
-import { bookStatusMap, getCategoryNameById, tradeStatusMap } from '../lib';
+import {
+  bookStatusMap,
+  getCategoryNameById,
+  tradeStatusMap,
+} from '../../../shared/lib';
 import * as S from './ListingCard.styles';
+import { ListingCardProps } from '@/entities/listing/model/types';
 
-interface ListingCardProps {
-  postId: number;
-  categoryId: number;
-  title: string;
-  tradeStatus: 'READY' | 'IN_PROGRESS' | 'COMPLETED';
-  sellPrice: number;
-  thumbnail: string;
-  bookStatus: 'BEST' | 'HIGH' | 'MEDIUM' | 'LOW';
-  createdAt: string;
-}
-
-const data: ListingCardProps = {
-  postId: 1,
-  categoryId: 10,
-  title: '자바 성능 튜닝 이야기 - 개발자가 반드시 알아야 할',
-  tradeStatus: 'READY',
-  sellPrice: 8000,
-  thumbnail:
-    'https://image.aladin.co.kr/product/7924/83/cover500/k542434036_1.jpg',
-  bookStatus: 'BEST',
-  createdAt: '2024-03-29T10:22:00',
-};
-
-const ListingCard = () => {
+const ListingCard = ({ data }: { data: ListingCardProps }) => {
   const categoryName = getCategoryNameById(data.categoryId);
 
   return (
