+++ src/features/listing/ui/ListingList.tsx
@@ -1,12 +1,15 @@
 'use client';
 import styled from 'styled-components';
 import ListingCard from './ListingCard';
+import { data } from '@/mocks/mockListingList';
 
 const ListingList = () => {
   return (
     <Container>
       판매글 리스트
-      <ListingCard />
+      {data.map((listing, idx) => (
+        <ListingCard key={listing.postId + idx} data={listing} />
+      ))}
     </Container>
   );
 };
