+++ src/mocks/mockListingList.ts
@@ -1,0 +1,15 @@
+import { ListingCardProps } from '@/entities/listing/model/types';
+
+export const data: ListingCardProps[] = [
+  {
+    postId: 1,
+    categoryId: 10,
+    title: '자바 성능 튜닝 이야기 - 개발자가 반드시 알아야 할',
+    tradeStatus: 'READY',
+    sellPrice: 8000,
+    thumbnail:
+      'https://image.aladin.co.kr/product/7924/83/cover500/k542434036_1.jpg',
+    bookStatus: 'BEST',
+    createdAt: '2024-03-29T10:22:00',
+  },
+];
