+++ src/shared/ui/index.ts
@@ -2,15 +2,6 @@ import Button from './Button';
 import Dropdown from './Dropdown';
 import Header from './Header';
 import InputGroup from './InputGroup';
-import ListingList from './ListingList';
 import SelectAreaSection from './SelectAreaSection';
 import CheckBox from './CheckBox';
-export {
-  Button,
-  CheckBox,
-  Dropdown,
-  Header,
-  InputGroup,
-  ListingList,
-  SelectAreaSection,
-};
+export { Button, CheckBox, Dropdown, Header, InputGroup, SelectAreaSection };
